Fix 9

+++ include/cell.h
@@ -17,7 +17,8 @@ class Cell {
   Ocean *ocean;
 
  public:
-  explicit Cell(Pair _p = {0, 0}, Ocean *_ocean = nullptr)
+  explicit Cell(Pair _p = {narrow_cast<int>(0), narrow_cast<int>(0)},
+                Ocean *_ocean = nullptr)
       : crd(_p), obj(nullptr), ocean(_ocean) {}
   Object *getObject() const;
   void setObject(Object *);
